Moved litmus-specific code into separate base class
