+++ test/FubarDev.WebDavServer.Tests/LitmusTolsen/Locks.cs
@@ -26,7 +26,7 @@ using Xunit;
 
 namespace FubarDev.WebDavServer.Tests.LitmusTolsen
 {
-    public class Locks : ServerTestsBase
+    public class Locks : LitmusTestsBase
     {
         private const string FooContent = @"This
 is
@@ -38,19 +38,12 @@ foo
 
 ";
 
-        private static readonly WebDavTimeoutHeaderValue _defaultTimeout =
-            WebDavTimeoutHeaderValue.CreateWebDavTimeout(TimeSpan.FromHours(1));
-
-        private static readonly LockScope _exclusive = LockScope.CreateExclusiveLockScope();
-
-        private static readonly LockType _write = LockType.CreateWriteLockType();
-
         private readonly string _dataPath = CreateDataPath();
 
         [Fact]
         public async Task UnmapLockRoot()
         {
-            await InitAsync(66, "unmap_lockroot", "lockcoll");
+            await InitAsync("locks", 66, "unmap_lockroot", "lockcoll");
 
             var secondClient = CreateClone("X-Litmus-Second");
 
@@ -157,99 +150,6 @@ foo
                 Guid.NewGuid().ToString("N"));
             Directory.CreateDirectory(tempRootPath);
             return tempRootPath;
-        }
-
-        private async Task InitAsync(int index, string name, params string[] rootPathParts)
-        {
-            // Initialize client with default values
-            Client.DefaultRequestHeaders.Add("X-Litmus", $"locks: {index} ({name})");
-            (await Client.MkcolAsync(new Uri("litmus", UriKind.Relative))).EnsureSuccessStatusCode();
-
-            // Create the root path
-            var baseAddress = new Uri(Client.BaseAddress!, new Uri("litmus/", UriKind.Relative));
-            foreach (var rootPathPart in rootPathParts)
-            {
-                var oldClient = Client;
-                Client = CreateClone();
-                Client.BaseAddress = baseAddress;
-                oldClient.Dispose();
-
-                (await Client.MkcolAsync(new Uri(rootPathPart, UriKind.Relative))).EnsureSuccessStatusCode();
-                baseAddress = new Uri(baseAddress, rootPathPart + "/");
-            }
-
-            {
-                var oldClient = Client;
-                Client = CreateClone();
-                Client.BaseAddress = baseAddress;
-                oldClient.Dispose();
-            }
-        }
-
-        private async Task<WebDavResponseMessage> LockAsync(string path, WebDavDepthHeaderValue depth)
-        {
-            var response = await Client.LockAsync(
-                path,
-                _defaultTimeout,
-                depth,
-                new LockInfo()
-                {
-                    LockScope = _exclusive,
-                    LockType = _write,
-                    OwnerHref = "litmus test suite",
-                });
-
-            // Add lock token as default header to the client.
-            if (response.IsSuccessStatusCode)
-            {
-                var lockToken = await WebDavHelper.GetLockTokenFromWebDavResponseMessage(response);
-                if (lockToken != null)
-                {
-                    Client.DefaultRequestHeaders.Add("If", lockToken.IfHeaderNoTagListFormat.ToString());
-                }
-            }
-
-            return response;
-        }
-
-        private async Task<WebDavResponseMessage> UnlockAsync(string path, LockToken lockToken)
-        {
-            // Remove the lock token from the default header.
-            var lockTokenHeader = lockToken.IfHeaderNoTagListFormat.ToString();
-            var headers = Client.DefaultRequestHeaders.GetValues("If")
-                .Where(x => x != lockTokenHeader)
-                .ToList();
-            Client.DefaultRequestHeaders.Remove("If");
-            if (headers.Count != 0)
-            {
-                Client.DefaultRequestHeaders.Add("If", headers);
-            }
-
-            // Perform the unlock. The lock token is automatically added to the request header.
-            var response = await Client.UnlockAsync(path, lockToken);
-
-            // Add the lock token to the header in case of an error.
-            if (!response.IsSuccessStatusCode)
-            {
-                Client.DefaultRequestHeaders.Add("If", lockTokenHeader);
-            }
-
-            return response;
-        }
-
-        private WebDavClient CreateClone(string litmusHeaderName = "X-Litmus")
-        {
-            var client = new WebDavClient(Server.CreateHandler())
-            {
-                BaseAddress = Client.BaseAddress,
-            };
-
-            if (Client.DefaultRequestHeaders.TryGetValues("X-Litmus", out var litmusHeader))
-            {
-                client.DefaultRequestHeaders.Add(litmusHeaderName, litmusHeader);
-            }
-
-            return client;
         }
     }
 }
